requirements.txt
loadenv
django-htmx
requests